All functional now!
Already changed to set the wts of the write into the publicKeyBlock that
we are writing at the moment. So when we call fs_write(something to
write, wts)
the publicKeyBlock will have <Hashes of ContentHashBlocks, signature,
wts>
Instead of the old timestamp mechanism that we were going to implement.

All functional now!